Data management: fix package names (#965)
