+++ extensions/api/data-management/contractagreement/src/main/resources/META-INF/services/org.eclipse.dataspaceconnector.spi.system.ServiceExtension
@@ -12,4 +12,4 @@
 #
 #
 
-org.eclipse.dataspaceconnector.api.datamanagement.contractdefinition.ContractAgreementApiExtension
+org.eclipse.dataspaceconnector.api.datamanagement.contractagreement.ContractAgreementApiExtension
