+++ extensions/api/data-management/transferprocess/src/main/resources/META-INF/services/org.eclipse.dataspaceconnector.spi.system.ServiceExtension
@@ -12,4 +12,4 @@
 #
 #
 
-org.eclipse.dataspaceconnector.api.datamanagement.contractdefinition.TransferProcessApiExtension
+org.eclipse.dataspaceconnector.api.datamanagement.transferprocess.TransferProcessApiExtension
